# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShopVidi Marketplace Core API
  version: 1.0.0
  description: |
    Marketplace Core 服务 HTTP 契约（阶段 1）。

    版本策略：URL 路径版本化，破坏性变更进入新主版本（/api/v2/...），
    向后兼容的新增（新字段、新端点）不升主版本。详见 conventions.md。

    统一约定：
    - 请求追踪头 `X-Request-Id`（UUID v4）。
    - 幂等键头 `Idempotency-Key`。
    - 所有响应 JSON，时间 RFC 3339 UTC，金额用 amount_minor（见 schemas/）。

    阶段 1 纳入 Product 资源（创建 / 查询 / 更新 / 发布 / 下架）与 Brand 资源
    （创建 / 查询 / 更新 / 删除），数据形状分别以 `schemas/product.schema.json`、
    `schemas/brand.schema.json` 为权威（批发价 / MOQ 真相在 Core）。
  contact:
    name: ShopVidi rebuild
  servers:
    - url: http://localhost:8080
      description: 本地开发
  tags:
    - name: health
      description: 存活与就绪检查
    - name: brands
      description: 品牌（商品归属，Core 主权，运营代维护）
    - name: products
      description: 商品（Core 业务主实体，批发价 / MOQ 真相）
paths:
  /health:
    get:
      tags: [health]
      summary: 服务存活检查
      operationId: getHealth
      responses:
        "200":
          description: 服务存活
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /ready:
    get:
      tags: [health]
      summary: 依赖就绪检查
      operationId: getReady
      responses:
        "200":
          description: 依赖就绪
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/health:
    get:
      tags: [health]
      summary: 版本化健康检查
      operationId: getHealthV1
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: 服务存活
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/v1/brands:
    get:
      tags: [brands]
      summary: 分页列出品牌
      operationId: listBrands
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: 品牌列表
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandList"
    post:
      tags: [brands]
      summary: 创建品牌（运营）
      description: |
        在 Core 创建品牌档案，初始状态为 `active`。品牌是商品的归属，
        商品创建依赖 `brand_id`。`name` 全局唯一。
      operationId: createBrand
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandCreate"
      responses:
        "201":
          description: 创建成功，返回完整品牌（status=active）
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/brands/{brand_id}:
    parameters:
      - $ref: "#/components/parameters/BrandId"
    get:
      tags: [brands]
      summary: 查询单个品牌
      operationId: getBrand
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: 品牌详情
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [brands]
      summary: 更新品牌
      description: |
        更新品牌名称 / 描述 / Logo URL / 状态。`name` 全局唯一，冲突返回 409；
        `status` 可取 `active` / `inactive`。可显式将可选字段置 `null` 以清空。
      operationId: updateBrand
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdate"
      responses:
        "200":
          description: 更新成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [brands]
      summary: 删除品牌（仅无商品时）
      description: |
        删除品牌档案。品牌下存在商品时返回 409，需先下架或改绑其商品，
        以保护商品 `brand_id` 外键引用。
      operationId: deleteBrand
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "204":
          description: 已删除
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/products:
    get:
      tags: [products]
      summary: 分页列出商品
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: 商品列表
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      tags: [products]
      summary: 创建商品草稿（运营 / Admin）
      description: |
        在 Core 创建商品，初始状态为 `draft`（不投影到 Woo / Odoo）。
        批发价与 MOQ 是 Core 真相，创建即写入。发布见 `POST /products/{id}/publish`。
      operationId: createProduct
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: 创建成功，返回完整商品（status=draft，external_ids 为空）
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/products/{product_id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [products]
      summary: 查询单个商品
      description: |
        返回完整商品（含批发价 `wholesale_price` 与 MOQ）。
        授权语义：未认证买家不可见批发价 / MOQ（由 ISSUE-0103 在 API 层强制）；
        已认证买家与运营可见。Woo 桥接插件以买家身份调用本接口渲染批发价 / MOQ。
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: 商品详情
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [products]
      summary: 更新商品
      description: |
        更新商品名称 / 品牌 / 批发价 / MOQ。SKU 一经发布不可修改。
        更新后触发 `catalog.product.updated`，Integration 同步 Woo / Odoo 投影。
      operationId: updateProduct
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: 更新成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/products/{product_id}/publish:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    post:
      tags: [products]
      summary: 发布商品（draft → published）
      description: |
        Core 将商品状态推进为 `published` 并发出 `catalog.product.published` 事件；
        Integration 消费事件后创建 Woo 投影与 Odoo 投影，并把 `woo_product_id` /
        `odoo_product_id` 写回 Core。投影为异步：响应返回时 `external_ids` 可能仍为空。
      operationId: publishProduct
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 已发布（status=published）
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/products/{product_id}/archive:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    post:
      tags: [products]
      summary: 下架商品（published → archived）
      description: |
        Core 将商品状态推进为 `archived` 并发出 `catalog.product.archived` 事件；
        Integration 在 Woo 隐藏、在 Odoo 停用对应投影。
      operationId: archiveProduct
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 已下架（status=archived）
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  schemas:
    Health:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [ok, degraded, down]
          description: 整体状态。
        checks:
          type: object
          description: 各依赖检查项，键为依赖名。
          additionalProperties:
            type: object
            required: [status]
            additionalProperties: false
            properties:
              status:
                type: string
                enum: [ok, degraded, down]
              latency_ms:
                type: integer
                description: 检查耗时（毫秒，可选）。
    Product:
      $ref: "../schemas/product.schema.json"
    Brand:
      $ref: "../schemas/brand.schema.json"
    BrandCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        logo_url:
          type: string
          maxLength: 2048
    BrandUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: ["string", "null"]
          maxLength: 2000
        logo_url:
          type: ["string", "null"]
          maxLength: 2048
        status:
          type: string
          enum: [active, inactive]
    BrandList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "../schemas/brand.schema.json"
        total:
          type: integer
          minimum: 0
          description: 满足过滤条件的总数。
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    ProductCreate:
      type: object
      additionalProperties: false
      required: [sku, name, brand_id, wholesale_price, moq]
      properties:
        sku:
          $ref: "../schemas/external-ids.schema.json#/$defs/Sku"
        name:
          type: string
          minLength: 1
          maxLength: 255
        brand_id:
          type: string
          format: uuid
        wholesale_price:
          $ref: "../schemas/money.schema.json"
        moq:
          type: integer
          minimum: 1
    ProductUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        brand_id:
          type: string
          format: uuid
        wholesale_price:
          $ref: "../schemas/money.schema.json"
        moq:
          type: integer
          minimum: 1
    ProductList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "../schemas/product.schema.json"
        total:
          type: integer
          minimum: 0
          description: 满足过滤条件的总数。
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: 机器可读错误码，如 validation_error、not_found、conflict、upstream_timeout。
            message:
              type: string
              description: 人类可读错误描述。
            details:
              type: array
              items:
                type: object
                required: [field, reason]
                additionalProperties: false
                properties:
                  field:
                    type: string
                  reason:
                    type: string
            request_id:
              type: string
              format: uuid
              description: 与 X-Request-Id 一致的请求追踪 ID。
  headers:
    XRequestId:
      description: 请求追踪 ID（UUID v4），由调用方生成，服务端记录并在响应回显。
      schema:
        type: string
        format: uuid
      example: "3f2a1c4e-8b7d-4a9f-9c2e-1d0f3a5b6c7d"
    IdempotencyKey:
      description: 幂等键，命名规则见 conventions.md。
      schema:
        type: string
        maxLength: 255
      example: "core.order.export.1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d"
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: 请求追踪 ID。缺失时服务端生成并在响应头回显。
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 幂等键。写操作（POST/PUT/PATCH）建议携带；同名同负载安全重试，同名不同负载返回 409。
      schema:
        type: string
        maxLength: 255
    Limit:
      name: limit
      in: query
      required: false
      description: 分页大小，默认 20，最大 100。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      description: 分页偏移，从 0 开始。
      schema:
        type: integer
        minimum: 0
        default: 0
    ProductId:
      name: product_id
      in: path
      required: true
      description: Core 商品主键（UUID v4）。
      schema:
        type: string
        format: uuid
    BrandId:
      name: brand_id
      in: path
      required: true
      description: Core 品牌主键（UUID v4）。
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: 请求不合法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: 幂等冲突或状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServiceUnavailable:
      description: 依赖未就绪
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"